requirements.txt
pytz
requests
apscheduler
gunicorn
django 
django-crispy-forms 
django-dynamic-preferences
dj_database_url
youtube-dl 
google-api-python-client 
google_auth_oauthlib 
oauth2client
psycopg2-binary
python-dateutil
Pillow
dependency-injector